refactor(testing): Composable benchmark architecture with self-documenting specs
Decompose the Kafka benchmark framework into independently reusable
building blocks behind a TriggerDriver interface. Each trigger type
(Kafka, Webhook, etc.) implements one driver file — the harnesses,
measurement, and reporting work unchanged.

Key changes:
- Extract shared types, workflow builder, execution sampler, diagnostics,
  and throughput measurement into utils/benchmark/
- Introduce TriggerDriver/TriggerHandle interfaces for trigger-agnostic
  harnesses
- Convert harnesses from test factories (registerXTests) to utility
  functions (runLoadTest/runThroughputTest) called inside test bodies
- Inline scenarios with descriptive test names — each spec is
  self-documenting with no file hopping needed
- Decouple Kafka from benchmark base config in playwright-projects.ts
